% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_rn.R
\name{cas_rn}
\alias{cas_rn}
\title{A Function}
\usage{
cas_rn(
  file = "",
  analyte_column = 1,
  db = "NIST",
  header = TRUE,
  browser = "chrome",
  chromever = "93.0.4577.63",
  port = sample(1000:9999, 1),
  extraCapabilities = list(chromeOptions = list(args = c("--headless",
    "--disable-gpu"))),
  verbose = F
)
}
\arguments{
\item{file}{The file}

\item{db}{database}

\item{header}{TRUE}

\item{browser}{TRYE}

\item{port}{random}

\item{extraCapabilities}{Later}

\item{verbose}{NOPE}

\item{chomever}{number}
}
\description{
This function allows you to webscrape NIST
}
\examples{
cas_rn("file.xlsx")
}
\keyword{cas_rn}
